4.21.2 Update plus light fixes
Updated project to use 4.21.2
Created LightRestoreHold BP - object that, when picked up, restores light over time. Also has a light in the middle that should scale on player light
Removed some old, unused BPs
Created wall that holds the Player Light bar instead of it being attached to the camera. This is nicer.